+++ packages/stylefire/src/css/transform-props.ts
@@ -1,7 +1,7 @@
 type Dictionary = { [key: string]: boolean };
 
 const axes = ['', 'X', 'Y', 'Z'];
-const order = ['scale', 'rotate', 'skew', 'transformPerspective'];
+const order = ['translate', 'scale', 'rotate', 'skew', 'transformPerspective'];
 
 export const transformProps = order.reduce(
   (acc, key) => {