+++ playground/stylefire.js
@@ -3,7 +3,7 @@ import styler from 'stylefire';
 
 const position = pos => ref => ref && styler(ref).set(pos);
 
-const Icon = ({ id, pos }) => (
+const Icon = ({ id, pos, style }) => (
   <svg
     width="641px"
     height="422px"
@@ -12,21 +12,6 @@ const Icon = ({ id, pos }) => (
     xmlns="http://www.w3.org/2000/svg"
     style={{ display: 'block' }}
   >
-    <defs>
-      <rect id="path-1" x="121" y="84" width="130" height="127" />
-      <mask
-        id="mask-2"
-        maskContentUnits="userSpaceOnUse"
-        maskUnits="objectBoundingBox"
-        x="-2"
-        y="-2"
-        width="134"
-        height="131"
-      >
-        <rect x="119" y="82" width="134" height="131" fill="white" />
-        <use xlinkHref="#path-1" fill="black" />
-      </mask>
-    </defs>
     <g id="Page-1" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
       <g id="Group" fillRule="nonzero">
         <rect
@@ -37,13 +22,17 @@ const Icon = ({ id, pos }) => (
           width="640"
           height="421"
         />
-        <use
+        <rect
           id="outline"
           stroke="#FF1C68"
           mask="url(#mask-2)"
           strokeWidth="4"
+          fill="white"
           strokeDasharray="10,10"
-          xlinkHref="#path-1"
+          x="120"
+          y="83"
+          width="102"
+          height="102"
         />
         <rect
           ref={position(pos)}
@@ -51,8 +40,9 @@ const Icon = ({ id, pos }) => (
           fill="#FF1C68"
           x="121"
           y="84"
-          width="130"
-          height="130"
+          width="100"
+          height="100"
+          style={style}
         />
       </g>
       <text x="20" y="20" fill="black">
@@ -66,16 +56,12 @@ export class SVG extends React.Component {
   render() {
     return (
       <Fragment>
-        <Icon pos={{ x: 100 }} />
-        <Icon pos={{ y: 100 }} />
+        <Icon pos={{ translateX: 100 }} />
+        <Icon pos={{ translateY: 100 }} />
         <Icon pos={{ rotate: 45 }} />
         <Icon pos={{ scale: 0.5 }} />
-        <Icon pos={{ x: 100, rotate: 45 }} />
-        <Icon pos={{ x: 100, y: 100, rotate: 45, scale: 0.5 }} />
-        <Icon pos={{ originX: 0, originY: 0, rotate: 45 }} />
-        <Icon pos={{ originX: '0%', originY: '0%', rotate: 45 }} />
-        <Icon pos={{ originX: '50%', originY: '50%', rotate: 45 }} />
-        <Icon pos={{ originX: '100%', originY: '100%', rotate: 45 }} />
+        <Icon pos={{ translateX: 100, rotate: 45, scale: 0.5 }} />
+        <Icon pos={{ rotate: 45, originX: 0, originY: 0 }} />
       </Fragment>
     );
   }